+++ src/CircuitBreaker.php
@@ -24,7 +24,8 @@ use Innmind\Immutable\{
 };
 
 /**
- * @psalm-import-type Errors from Transport
+ * @internal
+ * @psalm-import-type Errors from Implementation
  */
 final class CircuitBreaker implements Implementation
 {