+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Changed
 
 - Requires PHP `8.4`
+- `Innmind\HttpTransport\Transport` is now a final class, all previous implementations are now flagged as internal
 
 ### Removed